Update to pandas 2.3.3
Before, our `pandas` version was locked to v1.5.3 or lower, which came
out in Jan 2023 (almost 3 years ago!).

This commit updates our `pandas` version to 2.3.3 (Sep 2025) so that
we're using a more modern version of `pandas`.

Luckily the only backwards-incompatible change that has happened since
then was the removal of `df.append()`. I grepped the dsc10 private repo
from Spring 2025 and only found one use of `df.append()` in
`lectures/lec04/data-prep/get-it-done-data-prep.ipynb`. Since this was a
data preparation script and not student-facing, I decided to just remove
the `.append()` method on `bpd.DataFrame` and all of our test cases
still pass.

I also upgrade the build process to use `uv` instead of `setuptools`,
which generally considered the more modern way of doing things now.